+++ lizard_ui/static/lizard_ui/lizard_ui.js
@@ -340,8 +340,6 @@ $(document).ready(function() {
       return window.rightbarState = "opened";
     }
   });
-  // Disabled due to SSO login
-  /*
   $('.ui-login-link').click(function(e) {
     e.preventDefault();
     $('#login-modal').modal('toggle');
@@ -360,7 +358,6 @@ $(document).ready(function() {
     e.preventDefault();
     return handleLogin();
   });
-  */
 });
 
 function setup_appscreen_instant_load() {